♻️ refactor: rename enums to constants
- rename src/enums → src/constants (nothing there was actually a
  TS enum — DATA_ATTR/REGEX/BINDING_PROP/etc. are all `as const`
  values, so the name never matched the contents)
- update every import (relative and ~/ alias) plus
  scripts/sync-draggable-items.mjs
- update README.md/README.ko.md/AGENTS.md path references
- also fix AGENTS.md's path-alias table, which had scrambled/
  garbled target paths (e.g. "src/cemnnoopst" instead of
  "src/components") predating this change
